@@ -39,27 +39,28 @@
     "jws": "*",
     "lib-jitsi-meet": "jitsi/lib-jitsi-meet",
     "postis": "^2.2.0",
-    "react": "15.3.2",
-    "react-dom": "15.3.2",
-    "react-native": "0.37.0",
+    "react": "15.4.2",
+    "react-dom": "15.4.2",
+    "react-native": "0.42.3",
+    "react-native-background-timer": "1.0.0",      
     "react-native-loading-spinner-overlay": "^0.4.1",
     "react-native-navbar": "^1.5.4",
-    "react-native-permissions": "^0.2.5",
+    "react-native-permissions": "^0.2.7",
     "react-native-sound": "github:richmonkey/react-native-sound",
     "react-native-toast": "^1.0.1",
-    "react-native-vector-icons": "^2.0.3",
-    "react-native-webrtc": "github:richmonkey/react-native-webrtc",
-    "react-redux": "^4.4.6",
+    "react-native-vector-icons": "4.0.1",      
+    "react-native-webrtc": "jitsi/react-native-webrtc",
+    "react-redux": "5.0.4",
     "react-router": "^3.0.0",
     "react-router-redux": "^4.0.7",
-    "redux": "^3.5.2",
-    "redux-thunk": "^2.1.0",
+    "redux": "^3.6.0",
+    "redux-thunk": "^2.2.0",
     "retry": "0.6.1",
-    "strophe": "^1.2.2",
-    "strophejs-plugins": "^0.0.6",
-    "toastr": "^2.0.3",
+    "strophe": "1.2.4",
+    "strophejs-plugins": "0.0.7",
+    "toastr": "2.1.2",
     "url-polyfill": "github/url-polyfill",
-    "xmldom": "^0.1.22"
+    "xmldom": "^0.1.27"
   },
   "devDependencies": {
     "babel-core": "*",
